Add Output (cache-free GPIO), wire Peripheral into Context, comment all of it
Output (include/output.h, src/output.c) is a GPIO output Agent that
keeps no cached value: Set()/Get() always go straight to gpio_put()/
gpio_get(). Discussed at length before writing it - the alternative
(a cached field, refreshed either by SetOutputValue() alone or by a
periodic behave()) always ends up either relying on every writer
going through Set() forever to stay correct, or paying for a real
FreeRTOS task to poll a value that's cheap to read directly and that
only this program ever changes anyway. Going cache-free removes the
question entirely: there is only one source of truth (the pin itself),
so Output is created with timeout == 0 (no task, see Output_Behave)
and Output_Delete() now actually drives the pin low before freeing it,
instead of only clearing a soon-to-be-removed cached field.

Context gained a peripheral field (Peripheral*, set once in
NewContext() via GetPeripheral()) so any State can reach the board's
hardware without calling GetPeripheral() itself. Peripheral gained the
matching out1 field, alongside agent/led.

StateLoop_Run toggles out1 and reads it back every cycle as a manual
blink test - proving Get always agrees with Set with no cache in the
middle - and stays in STATE_LOOP on purpose for now (its transition to
STATE_SUB is commented out) while this is being exercised.

Also documented every public type/function touched above that was
still missing its @brief (Output, BlinkLed, Peripheral, and the new
Context.peripheral field), consistent with the rest of the codebase.

Dropped src/thyonei.c from the CMakeLists.txt build: it's still an
empty stub, not ready yet, and neither it nor its header are part of
this commit.

+++ include/blinkled.h
@@ -4,12 +4,33 @@
 #include "agent.h"
 #include "pico/stdlib.h"
 
+/**
+ * @brief A blinking LED, as an Agent extension: base (Agent) plus the
+ * GPIO pin it drives and its current on/off state. Embeds Agent as its
+ * first field (base) so a BlinkLed* can be upcast to Agent* (e.g.
+ * &led->base) and, from inside its own behave()/delete(), downcast
+ * back with (BlinkLed *)self - the same first-member idiom SubState
+ * uses on State.
+ */
 typedef struct {
-    Agent base;  
-    uint pin;
+    Agent base;
+    uint8_t pin;
     bool state;
 } BlinkLed;
 
-BlinkLed *NewBlinkLed(char *name, int timeout, uint32_t uxStackDepth, UBaseType_t uxPriority, uint pin);
+/**
+ * @brief Allocates a new BlinkLed, configures its GPIO as an output and
+ * registers it via Agent_Init() - its own behave() toggles the pin
+ * every timeout ms, so unlike Output it does need a real task
+ * (timeout > 0).
+ *
+ * @param name Name of the task/agent (copied internally, see Agent_Init).
+ * @param timeout Period in ms between toggles.
+ * @param uxStackDepth Task stack depth, in words.
+ * @param uxPriority FreeRTOS task priority (0 .. configMAX_PRIORITIES-1).
+ * @param pin GPIO pin number this LED is wired to.
+ * @return BlinkLed* Newly heap-allocated, already-blinking BlinkLed.
+ */
+BlinkLed *NewBlinkLed(char *name, int timeout, uint32_t uxStackDepth, UBaseType_t uxPriority, uint8_t pin);
 
 #endif
